Fix bad rounding using floats instead of doubles

+++ Android/EasyBudget/app/src/main/java/com/benoitletondor/easybudgetapp/model/Expense.java
@@ -41,7 +41,7 @@ public class Expense implements Serializable
     /**
      * Amount of this expense (can be < 0).
      */
-    private float     amount;
+    private double  amount;
     /**
      * Date of the expense
      */
@@ -59,7 +59,7 @@ public class Expense implements Serializable
      * @param amount
      * @param date
      */
-    public Expense(@NonNull String title, float amount, @NonNull Date date)
+    public Expense(@NonNull String title, double amount, @NonNull Date date)
     {
         this(null, title, amount, date, null);
     }
@@ -71,7 +71,7 @@ public class Expense implements Serializable
      * @param date
      * @param monthlyId
      */
-    public Expense(@NonNull String title, float amount, @NonNull Date date, Long monthlyId)
+    public Expense(@NonNull String title, double amount, @NonNull Date date, Long monthlyId)
     {
         this(null, title, amount, date, monthlyId);
     }
@@ -84,7 +84,7 @@ public class Expense implements Serializable
      * @param date
      * @param monthlyId
      */
-    public Expense(Long id, @NonNull String title, float amount, @NonNull Date date, Long monthlyId)
+    public Expense(Long id, @NonNull String title, double amount, @NonNull Date date, Long monthlyId)
     {
         this.id = id;
 
@@ -154,12 +154,12 @@ public class Expense implements Serializable
         this.date = DateHelper.cleanDate(date);
     }
 
-    public float getAmount()
+    public double getAmount()
     {
         return amount;
     }
 
-    public void setAmount(float amount)
+    public void setAmount(double amount)
     {
         this.amount = amount;
     }
